Add a overall files count and uploaded chunk function
I am the maintainer of the Flowupload app for Nextcloud which uses ng-flow extensively. Recently @TessyPowder added two functions to ng-flow which adds a counter for all files and a function to count already uploaded chunks. I would like to bring these changes to upstream so other projects can use them as well and so they are in the right place.

The original PR on my repository can be found here:

https://github.com/e-alfred/flowupload/pull/57